+++ src/components/Game.tsx
@@ -201,33 +201,33 @@ const Game: React.FC<IGame> = ({ selectedTeam }) => {
       (r: any) =>
         r.data.ix === "3u32.private" &&
         r.data.challenger.replace(".private", "") ===
-        currentGame.gameNotification.recordData.challenger_address &&
+          currentGame.gameNotification.recordData.challenger_address &&
         r.data.staker.replace(".private", "") ===
-        currentGame.gameNotification.recordData.challenger_address
+          currentGame.gameNotification.recordData.challenger_address
     );
     const piece_claim_challenger = msPuzzleRecords.find(
       (r: any) =>
         r.data.ix === "6u32.private" &&
         r.data.challenger.replace(".private", "") ===
-        currentGame.gameNotification.recordData.challenger_address &&
+          currentGame.gameNotification.recordData.challenger_address &&
         r.data.claimer.replace(".private", "") ===
-        currentGame.gameNotification.recordData.challenger_address
+          currentGame.gameNotification.recordData.challenger_address
     );
     const piece_stake_opponent = msPuzzleRecords.find(
       (r) =>
         r.data.ix === "3u32.private" &&
         r.data.opponent.replace(".private", "") ===
-        currentGame.gameNotification.recordData.opponent_address &&
+          currentGame.gameNotification.recordData.opponent_address &&
         r.data.staker.replace(".private", "") ===
-        currentGame.gameNotification.recordData.opponent_address
+          currentGame.gameNotification.recordData.opponent_address
     );
     const piece_claim_opponent = msPuzzleRecords.find(
       (r) =>
         r.data.ix === "6u32.private" &&
         r.data.opponent.replace(".private", "") ===
-        currentGame.gameNotification.recordData.opponent_address &&
+          currentGame.gameNotification.recordData.opponent_address &&
         r.data.claimer.replace(".private", "") ===
-        currentGame.gameNotification.recordData.opponent_address
+          currentGame.gameNotification.recordData.opponent_address
     );
 
     console.log("msGameRecords[0]", msGameRecords[0]);
@@ -396,7 +396,7 @@ const Game: React.FC<IGame> = ({ selectedTeam }) => {
           challenger_nonce: messageToSign, /// todo - make this random
           challenger_answer: "[" + activePlayerIds.toString() + "]",
           game_multisig_seed,
-          uuid: "12423423field", // uuidv4(), TODO generate random uuid
+          uuid: uuidv4(), // uuidv4(), TODO generate random uuid
         };
         console.log(
           "🚀 ~ createProposeGameEvent ~ proposalInputs:",
